fix: Resolve HackerPack syntax issue
This came from a rebase conflict that wasn't caught until after a merge
